requirements.txt
flask
pyjwt==2.0.0
datetime
uuid
Flask-SQLAlchemy
Flask-Migrate
flask-script==2.0.5
marshmallow
flask-swagger
flask-blueprint